Refactor driver to be runtime configurable
Update the atmel-rf-driver to be runtime configurable rather than
setup at link time. Remove most public functions to prevent
conflicts with other rf drivers which may be used at the same time.

More detailed list of changes:
-Remove all static pin declarations and configurations
-Covert I2C mac driver into a class which has the SCL and SDA pin as
    parameters
-Combine driverAtmelRFInterface.cpp, driverAtmelRFInterface.h and
    driverRFPhy.h into NanostackRfPhyAtmel.cpp
-Remove dead and unused functions
-Create NanostackRfPhyAtmel class which implements NanostackRfPhy
    and takes necessary pins as constructor parameters
-Make all functions in NanostackRfPhyAtmel.cpp private and only
    expose NanostackRfPhyAtmel publicly
-Move all AT86RF specific defines into AT86RFReg.h
