--- mbed_lib.json
@@ -1,14 +1,0 @@
-{
-    "name": "atmel-rf",
-    "config": {
-        "spi-mosi": null,
-        "spi-miso": null,
-        "spi-sclk": null,
-        "spi-cs": null,
-        "spi-rst": null,
-        "spi-slp": null,
-        "spi-irq": null,
-        "i2c-sda": null, 
-        "i2c-scl": null 
-    }
-}
