+++ source/at24mac.h
@@ -16,102 +16,48 @@
 #ifndef AT24MAC_H
 #define AT24MAC_H
 
+#include "PinNames.h"
+#include "I2C.h"
+
 /*
  * AT24MAC drivers.
  *
  * This is a EEPROM chip designed to contain factory programmed read-only EUI-64 or EUI-48,
  * a 128bit serial number and some user programmable EEPROM.
  *
- * AT24MAC602 contains EUI-64, use at24mac_read_eui64()
- * AT24MAC402 contains EUI-64, use at24mac_read_eui48()
+ * AT24MAC602 contains EUI-64, use read_eui64()
+ * AT24MAC402 contains EUI-64, use read_eui48()
  *
  * NOTE: You cannot use both EUI-64 and EUI-48. Chip contains only one of those.
  */
 
-/* 
- * Default interface pins definition. 
- * Can be redefined using a config.json file or the config section of the used target
- * using the following structure and changing the "i2c" section.
- *
- * "atmel": {
- *    "rf": {
- *      "spi": {
- *        "mosi": "",
- *        "miso": "",
- *        "sclk": "",
- *        "cs": "",
- *        "rst": "",
- *        "slp": "",
- *        "irq": ""
- *      },
- *      "i2c": {
- *    	  "sda": "", 
- *    	  "scl": "" 
- *      }
- *    }
- *  }
- */
+class AT24Mac {
+public:
+    AT24Mac(PinName sda, PinName scl);
 
-/* mbed-cli configuration */
-#ifdef MBED_CONF_ATMEL_RF_I2C_SDA
-#define PIN_I2C_SDA MBED_CONF_ATMEL_RF_I2C_SDA
-#endif
-#ifdef MBED_CONF_ATMEL_RF_I2C_SCL
-#define PIN_I2C_SCL MBED_CONF_ATMEL_RF_I2C_SCL
-#endif
+    /**
+     * Read unique serial number from chip.
+     * \param buf pointer to write serial number to. Must have space for 16 bytes.
+     * \return zero on success, negative number on failure
+     */
+    int read_serial(void *buf);
 
-/* Yotta configuration */
-#ifdef YOTTA_CFG_ATMEL_RF_I2C_SDA
-#define PIN_I2C_SDA YOTTA_CFG_ATMEL_RF_I2C_SDA
-#endif
-#ifdef YOTTA_CFG_ATMEL_RF_I2C_SCL
-#define PIN_I2C_SCL YOTTA_CFG_ATMEL_RF_I2C_SCL
-#endif
+    /**
+     * Read EUI-64 from chip.
+     * \param buf pointer to write EUI-64 to. Must have space for 8 bytes.
+     * \return zero on success, negative number on failure
+     */
+    int read_eui64(void *buf);
 
-/* Defaults for Arduino form factor, or Yotta */
-#if defined TARGET_FF_ARDUINO || defined YOTTA_CFG
-#ifndef PIN_I2C_SDA
-#define PIN_I2C_SDA D14
-#endif
-#ifndef PIN_I2C_SCL
-#define PIN_I2C_SCL D15
-#endif
-#endif //TARGET_FF_ARDUINO || YOTTA_CFG
+    /**
+     * Read EUI-48 from chip.
+     * \param buf pointer to write EUI-48 to. Must have space for 6 bytes.
+     * \return zero on success, negative number on failure
+     */
+    int read_eui48(void *buf);
 
-/* Final fallback defines so we at least compile */
-#ifndef PIN_I2C_SDA
-#define PIN_I2C_SDA NC
-#endif
-#ifndef PIN_I2C_SCL
-#define PIN_I2C_SCL NC
-#endif
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/**
- * Read uniqua serial number from chip.
- * \param buf pointer to write serial number to. Must have space for 16 bytes.
- * \return zero on success, negative number on failure
- */
-int at24mac_read_serial(void *buf);
-
-/**
- * Read EUI-64 from chip.
- * \param buf pointer to write EUI-64 to. Must have space for 8 bytes.
- * \return zero on success, negative number on failure
- */
-int at24mac_read_eui64(void *buf);
-/**
- * Read EUI-48 from chip.
- * \param buf pointer to write EUI-48 to. Must have space for 6 bytes.
- * \return zero on success, negative number on failure
- */
-int at24mac_read_eui48(void *buf);
-
-#ifdef __cplusplus
-}
-#endif
+private:
+    mbed::I2C _i2c;
+};
 
 #endif /* AT24MAC_H */
